requirements.txt
# Outils de build
setuptools>=65.0.0
wheel>=0.38.0

# Framework web
Flask==3.0.0
Flask-Cors==4.0.0
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.3

# Traitement d'images et IA
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0.74
Pillow>=10.0.0,<11.0.0
tensorflow>=2.13.0,<2.16.0
scikit-learn>=1.3.0

# Stéganographie et hashing
stegano>=0.9.4
imagehash>=4.3.1

# Base de données
SQLAlchemy>=2.0.23

# Utilitaires
python-dotenv>=1.0.0
Werkzeug>=3.0.0

# Développement et tests
pytest>=7.4.0
pytest-flask>=1.3.0
pytest-cov>=4.1.0

# Sécurité
bcrypt>=4.0.1